+++ src/main/java/org/fulib/classmodel/Attribute.java
@@ -1,9 +1,19 @@
 package org.fulib.classmodel;
 
+import org.antlr.v4.runtime.CharStream;
+import org.antlr.v4.runtime.CharStreams;
+import org.antlr.v4.runtime.CommonTokenStream;
 import org.fulib.builder.Type;
+import org.fulib.parser.FragmentMapBuilder;
+import org.fulib.parser.FulibClassLexer;
+import org.fulib.parser.FulibClassParser;
+import org.fulib.util.Validator;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 
 public class Attribute
@@ -16,6 +26,10 @@ public class Attribute
    public static final String PROPERTY_collectionType = "collectionType";
    public static final String PROPERTY_initialization = "initialization";
    public static final String PROPERTY_propertyStyle = "propertyStyle";
+   /** @since 1.3 */
+   public static final String PROPERTY_description = "description";
+   /** @since 1.3 */
+   public static final String PROPERTY_since = "since";
    public static final String PROPERTY_modified = "modified";
    public static final String PROPERTY_clazz = "clazz";
 
@@ -29,7 +43,11 @@ public class Attribute
    private CollectionType collectionType;
    private String initialization;
    private String propertyStyle;
+   private String description;
+   private String since;
    private boolean modified;
+
+   private String typeSignature;
 
    // =============== Properties ===============
 
@@ -94,7 +112,7 @@ public class Attribute
       return this.type;
    }
 
-   public Attribute setType(String value)
+   public Attribute setType(String value) // no fulib
    {
       if (Objects.equals(value, this.type))
       {
@@ -103,8 +121,38 @@ public class Attribute
 
       final String oldValue = this.type;
       this.type = value;
+      this.typeSignature = buildTypeSignature(value);
       this.firePropertyChange(PROPERTY_type, oldValue, value);
       return this;
+   }
+
+   private static String buildTypeSignature(String type)
+   {
+      if (Validator.isSimpleName(type))
+      {
+         // fast-path - if the type is just an identifier (no generics, no annotations, ...),
+         // we can skip all the parser overhead because the signature will be same as the type anyway.
+         return type;
+      }
+
+      final CharStream input = CharStreams.fromString(type);
+      final FulibClassLexer lexer = new FulibClassLexer(input);
+      final FulibClassParser parser = new FulibClassParser(new CommonTokenStream(lexer));
+      final FulibClassParser.TypeContext typeCtx = parser.type();
+      return FragmentMapBuilder.getTypeSignature(typeCtx);
+   }
+
+   /**
+    * @return the signature of this attribute's {@linkplain #getType() type}
+    *
+    * @deprecated for internal use only
+    *
+    * @since 1.2.2
+    */
+   @Deprecated
+   public String getTypeSignature()
+   {
+      return this.typeSignature;
    }
 
    /**
@@ -119,9 +167,9 @@ public class Attribute
 
    /**
     * @param value
-    *    the new collection type
+    *    the collection type
     *
-    * @return this instance, to allow method chaining
+    * @return this
     *
     * @since 1.2
     */
@@ -167,7 +215,8 @@ public class Attribute
    }
 
    /**
-    * @return the property style of this attribute
+    * @return the property style.
+    * Currently, only {@link Type#POJO}, {@link Type#BEAN} and {@link Type#JAVA_FX} are supported.
     */
    public String getPropertyStyle()
    {
@@ -176,10 +225,10 @@ public class Attribute
 
    /**
     * @param value
-    *    the property style to use for this attribute.
+    *    the property style.
     *    Currently, only {@link Type#POJO}, {@link Type#BEAN} and {@link Type#JAVA_FX} are supported.
     *
-    * @return this instance, to allow method chaining
+    * @return this
     */
    public Attribute setPropertyStyle(String value)
    {
@@ -195,6 +244,80 @@ public class Attribute
    }
 
    /**
+    * @return the description of this attribute, used for generating JavaDocs
+    *
+    * @since 1.3
+    */
+   public String getDescription()
+   {
+      return this.description;
+   }
+
+   /**
+    * @param value
+    *    the description of this attribute, used for generating JavaDocs
+    *
+    * @return this
+    *
+    * @since 1.3
+    */
+   public Attribute setDescription(String value)
+   {
+      if (Objects.equals(value, this.description))
+      {
+         return this;
+      }
+
+      final String oldValue = this.description;
+      this.description = value;
+      this.firePropertyChange(PROPERTY_description, oldValue, value);
+      return this;
+   }
+
+   /**
+    * @return the lines of the description of this attribute, used for generating JavaDocs
+    *
+    * @since 1.3
+    * @deprecated for internal use only
+    */
+   @Deprecated
+   public List<String> getDescriptionLines()
+   {
+      return this.getDescription() == null ? Collections.emptyList() : Arrays.asList(this.getDescription().split("\n"));
+   }
+
+   /**
+    * @return the version when this attribute was introduced, used for generating JavaDocs
+    *
+    * @since 1.3
+    */
+   public String getSince()
+   {
+      return this.since;
+   }
+
+   /**
+    * @param value
+    *    the version when this attribute was introduced, used for generating JavaDocs
+    *
+    * @return this
+    *
+    * @since 1.3
+    */
+   public Attribute setSince(String value)
+   {
+      if (Objects.equals(value, this.since))
+      {
+         return this;
+      }
+
+      final String oldValue = this.since;
+      this.since = value;
+      this.firePropertyChange(PROPERTY_since, oldValue, value);
+      return this;
+   }
+
+   /**
     * @return a boolean indicating whether this attribute was modified. For internal use only.
     */
    public boolean getModified()
@@ -206,7 +329,7 @@ public class Attribute
     * @param value
     *    a boolean indicating whether this attribute was modified. For internal use only.
     *
-    * @return this instance, to allow method chaining
+    * @return this
     */
    public Attribute setModified(boolean value)
    {
@@ -295,6 +418,8 @@ public class Attribute
       result.append(' ').append(this.getType());
       result.append(' ').append(this.getInitialization());
       result.append(' ').append(this.getPropertyStyle());
+      result.append(' ').append(this.getDescription());
+      result.append(' ').append(this.getSince());
       return result.substring(1);
    }
 }