+++ src/main/java/org/fulib/classmodel/CodeFragment.java
@@ -1,7 +1,5 @@
 package org.fulib.classmodel;
 
-import java.beans.PropertyChangeListener;
-import java.beans.PropertyChangeSupport;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.Objects;
@@ -16,8 +14,6 @@ public class CodeFragment extends Fragment
    public static final String PROPERTY_text = "text";
 
    // =============== Fields ===============
-
-   protected PropertyChangeSupport listeners;
 
    private String text;
 
@@ -58,60 +54,6 @@ public class CodeFragment extends Fragment
    public void write(Writer writer) throws IOException
    {
       writer.write(this.getText());
-   }
-
-   public boolean addPropertyChangeListener(PropertyChangeListener listener)
-   {
-      if (this.listeners == null)
-      {
-         this.listeners = new PropertyChangeSupport(this);
-      }
-      this.listeners.addPropertyChangeListener(listener);
-      return true;
-   }
-
-   public boolean addPropertyChangeListener(String propertyName, PropertyChangeListener listener)
-   {
-      if (this.listeners == null)
-      {
-         this.listeners = new PropertyChangeSupport(this);
-      }
-      this.listeners.addPropertyChangeListener(propertyName, listener);
-      return true;
-   }
-
-   public boolean removePropertyChangeListener(PropertyChangeListener listener)
-   {
-      if (this.listeners != null)
-      {
-         this.listeners.removePropertyChangeListener(listener);
-      }
-      return true;
-   }
-
-   public boolean removePropertyChangeListener(String propertyName, PropertyChangeListener listener)
-   {
-      if (this.listeners != null)
-      {
-         this.listeners.removePropertyChangeListener(propertyName, listener);
-      }
-      return true;
-   }
-
-   public boolean firePropertyChange(String propertyName, Object oldValue, Object newValue)
-   {
-      if (this.listeners != null)
-      {
-         this.listeners.firePropertyChange(propertyName, oldValue, newValue);
-         return true;
-      }
-      return false;
-   }
-
-   @Override
-   public void removeYou()
-   {
-      super.removeYou();
    }
 
    @Override